requirements.txt
appdirs==1.4.3
coverage==4.3.4
Django==1.10
factory-boy==2.8.1
Faker==0.7.11
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
